backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-socketio==5.11.1
pydantic==2.6.1
pytest==8.0.0
httpx==0.26.0
requests==2.31.0
